net/dpdk: upgrade to dpdk-19.05
- Move to new memory subsystem
  https://doc.dpdk.org/guides/rel_notes/release_18_05.html
  * Leverage VFIO+IOMMU to use IOVA, not PA, for DMA.
  * Direct map(IOVA=VA) each core's whole heap space at startup.

- Move to new offloads API
  DPDK ethdev offloads API has changed since:
    commit ce17eddefc20 ("ethdev: introduce Rx queue offloads API")
    commit cba7f53b717d ("ethdev: introduce Tx queue offloads API")

- Other changes
  * Drop explicit enabling CRC strip offload. From 18.11 release notes:
      The default behavior of CRC strip offload has changed in this
      release. Without any specific Rx offload flag, default behavior
      by a PMD is now to strip CRC.
  * Replace PKT_RX_VLAN_PKT with PKT_RX_VLAN. From 17.11 release notes:
      The mbuf flags PKT_RX_VLAN_PKT and PKT_RX_QINQ_PKT have been
      removed since their behavior was not properly described.
      Two mbuf flags have been added to indicate that the VLAN identifier
      has been saved in in the mbuf structure. For instance:
        If VLAN is not stripped and TCI is saved: PKT_RX_VLAN
        If VLAN is stripped and TCI is saved: PKT_RX_VLAN | PKT_RX_VLAN_STRIPPED
  * Set RSS offload rules per dev_info
      commit aa1a6d87f15d (ethdev: force RSS offload rules again)
  * Move xstats initialization after port started to fix errors on
    mellanox nics.
  * Update port id to 16 bits per 17.11 changes.
  * Replace deprecated function rte_eth_dev_count() with
    rte_eth_dev_count_avail().
  * Update DPDK build scripts and configs. Add "--whole-archive" when
    generating DPDK enabled applications.
  * Drop frag0 cross page checking as IOVA is always continuous.

Message-Id: <20190619054446.7674-1-yibo.cai@arm.com>

+++ include/seastar/net/dpdk.hh
@@ -31,8 +31,8 @@
 namespace seastar {
 
 std::unique_ptr<net::device> create_dpdk_net_device(
-                                    uint8_t port_idx = 0,
-                                    uint8_t num_queues = 1,
+                                    uint16_t port_idx = 0,
+                                    uint16_t num_queues = 1,
                                     bool use_lro = true,
                                     bool enable_fc = true);
